+++ src/main/java/com/d108/project/domain/businessArea/service/BusinessAreaService.java
@@ -5,5 +5,5 @@ import com.d108.project.domain.businessArea.dto.BusinessAreaListDto;
 import java.util.List;
 
 public interface BusinessAreaService {
-    BusinessAreaListDto getBusinessAreaList(List<Integer> businessAreaIds);
+    BusinessAreaListDto getBusinessAreaList(List<Long> businessAreaIds);
 }
