+++ src/main/java/com/d108/project/domain/favorite/favoriteBusinessArea/repository/FavoriteBusinessAreaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FavoriteBusinessAreaRepository extends JpaRepository<FavoriteBusinessArea, Integer> {
+public interface FavoriteBusinessAreaRepository extends JpaRepository<FavoriteBusinessArea, Long> {
     List<FavoriteBusinessArea> findAllByMember(Member member);
 }
