+++ src/main/java/com/d108/project/domain/franchise/repository/FranchiseRepository.java
@@ -3,5 +3,5 @@ package com.d108.project.domain.franchise.repository;
 import com.d108.project.domain.franchise.entity.Franchise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FranchiseRepository extends JpaRepository<Franchise, Integer> {
+public interface FranchiseRepository extends JpaRepository<Franchise, Long> {
 }
